+++ dutch/app/lib/turn-state.js
@@ -65,6 +65,8 @@ function createTurnState(deps) {
       a.card.id !== b.card.id
     ) {
       [a.player.cards[a.index], b.player.cards[b.index]] = [b.player.cards[b.index], a.player.cards[a.index]];
+      deps.markHandCardChanged(a.player.id, b.card.id);
+      deps.markHandCardChanged(b.player.id, a.card.id);
       deps.moveSlotMemoryForAllBots(a.player.id, a.index, b.player.id, b.index, 'Jack swap');
       if (deps.moveHumanKnowledgeForAllBots) {
         deps.moveHumanKnowledgeForAllBots(a.player.id, a.index, b.player.id, b.index, actorId);